# Slightly faster fast BLANC