A very minor speed squeezing tweak