@LoicGrobol 2020-10-28
Speedtest on my laptop with no particular care for isolation:

On master:

```text
Testing metrics speed: 100 calls, best of 5
metric  total   per call
MUC     9.520737202023156       0.09520737202023156
B³      32.07567617803579       0.32075676178035795
CEAF_m  11.991489637992345      0.11991489637992345
CEAF_e  16.462463992007542      0.16462463992007542
BLANC   1.0065736430115066      0.010065736430115067
Testing BLANC speedup: 100 calls, best of 5
Slow:   67.37697147001745 s (0.6737697147001745 s/call)
Fast:   1.0244948230101727 s (0.010244948230101726 s/call)
Speedup: ×65.76604386545391
```

This PR:

```text
Testing metrics speed: 100 calls, best of 5
metric  total   per call
MUC     9.854284832952544       0.09854284832952544
B³      33.92382380599156       0.3392382380599156
CEAF_m  12.061643913970329      0.12061643913970328
CEAF_e  15.983911918010563      0.15983911918010563
BLANC   0.9605731989722699      0.009605731989722699
Testing BLANC speedup: 100 calls, best of 5
Slow:   65.14477001497289 s (0.651447700149729 s/call)
Fast:   0.968178202980198 s (0.00968178202980198 s/call)
Speedup: ×67.28592919614127
```